pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "flask-wtf>=1.2.2",
    "wtforms>=3.2.1",
    "pillow>=11.3.0",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "pymysql>=1.1.1",
    "flask-dance>=7.1.0",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
]
